feat: improve rate display format in output message
Update the print statement to clarify which rate corresponds to GdLS and LS detectors by explicitly labeling them in the output. Also reorganize the unit notation to be more standard by moving "[Hz]" to the end of the line. This makes the output more readable and self-explanatory for users analyzing the rate data.